requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
python-dotenv==1.0.1
python-multipart==0.0.20
google-generativeai==0.8.4
python-docx==1.1.2
pydantic==2.10.4
pydantic-settings==2.7.1
aiofiles==24.1.0
